renamed app to businness logic
