+++ business_logic/tests/test_models.py
@@ -2,8 +2,8 @@ from django.test import TestCase
 
 from django.contrib.auth.models import User
 
-from backend.models import Bucket
-from backend.models import Goal
+from business_logic.models import Bucket
+from business_logic.models import Goal
 
 class BucketTestCase(TestCase):
     def setUp(self):
